Struct games1::ScoreMethodsBuilder

pub struct ScoreMethodsBuilder<'a, C, NC, A> where NC: 'a, C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on score resources. It is not used directly, but through the Games hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate games1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Games::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `get(...)`, `list(...)`, `list_window(...)`, `submit(...)` and `submit_multiple(...)`
// to build up your call.
let rb = hub.scores();

Methods

impl<'a, C, NC, A> ScoreMethodsBuilder<'a, C, NC, A>

fn submit(&self, leaderboard_id: &str, score: &str) -> ScoreSubmitMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Submits a score to the specified leaderboard.

fn list(&self, leaderboard_id: &str, collection: &str, time_span: &str) -> ScoreListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists the scores in a leaderboard, starting from the top.

fn get(&self, player_id: &str, leaderboard_id: &str, time_span: &str) -> ScoreGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Get high scores, and optionally ranks, in leaderboards for the currently authenticated player. For a specific time span, leaderboardId can be set to ALL to retrieve data for all leaderboards in a given time span. NOTE: You cannot ask for 'ALL' leaderboards and 'ALL' timeSpans in the same request; only one parameter may be set to 'ALL'.

fn submit_multiple(&self, request: &PlayerScoreSubmissionList) -> ScoreSubmitMultipleMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Submits multiple scores to leaderboards.

fn list_window(&self, leaderboard_id: &str, collection: &str, time_span: &str) -> ScoreListWindowMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists the scores in a leaderboard around (and including) a player's score.