Struct games1::TurnBasedMatcheMethodsBuilder

pub struct TurnBasedMatcheMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on turnBasedMatche resources. It is not used directly, but through the Games hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate games1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use games1::Games;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Games::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `cancel(...)`, `create(...)`, `decline(...)`, `dismiss(...)`, `finish(...)`, `get(...)`, `join(...)`, `leave(...)`, `leave_turn(...)`, `list(...)`, `rematch(...)`, `sync(...)` and `take_turn(...)`
// to build up your call.
let rb = hub.turn_based_matches();

Methods

impl<'a, C, NC, A> TurnBasedMatcheMethodsBuilder<'a, C, NC, A>

fn dismiss(&self, match_id: &str) -> TurnBasedMatcheDismisMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Dismiss a turn-based match from the match list. The match will no longer show up in the list and will not generate notifications.

fn sync(&self) -> TurnBasedMatcheSyncMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns turn-based matches the player is or was involved in that changed since the last sync call, with the least recent changes coming first. Matches that should be removed from the local cache will have a status of MATCH_DELETED.

fn decline(&self, match_id: &str) -> TurnBasedMatcheDeclineMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Decline an invitation to play a turn-based match.

fn get(&self, match_id: &str) -> TurnBasedMatcheGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Get the data for a turn-based match.

fn create(&self, request: &TurnBasedMatchCreateRequest) -> TurnBasedMatcheCreateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Create a turn-based match.

fn join(&self, match_id: &str) -> TurnBasedMatcheJoinMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Join a turn-based match.

fn leave_turn(&self, match_id: &str, match_version: i32) -> TurnBasedMatcheLeaveTurnMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Leave a turn-based match during the current player's turn, without canceling the match.

fn cancel(&self, match_id: &str) -> TurnBasedMatcheCancelMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Cancel a turn-based match.

fn finish(&self, request: &TurnBasedMatchResults, match_id: &str) -> TurnBasedMatcheFinishMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Finish a turn-based match. Each player should make this call once, after all results are in. Only the player whose turn it is may make the first call to Finish, and can pass in the final match state.

fn leave(&self, match_id: &str) -> TurnBasedMatcheLeaveMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Leave a turn-based match when it is not the current player's turn, without canceling the match.

fn list(&self) -> TurnBasedMatcheListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns turn-based matches the player is or was involved in.

fn take_turn(&self, request: &TurnBasedMatchTurn, match_id: &str) -> TurnBasedMatcheTakeTurnMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Commit the results of a player turn.

fn rematch(&self, match_id: &str) -> TurnBasedMatcheRematchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Create a rematch of a match that was previously completed, with the same participants. This can be called by only one player on a match still in their list; the player must have called Finish first. Returns the newly created match; it will be the caller's turn.